cogserv/requirements.txt
azure_storage==0.37.0
Django==4.0.4
msrest==0.6.21
Pillow==9.1.1
requests==2.27.1
